tests/requirements-test.txt
# Include base requirements
-r ../requirements.txt

# Test frameworks
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Browser automation (for @pytest.mark.browser tests)
selenium>=4.0.0

# Workflow execution (for @pytest.mark.workflow tests)
websocket-client>=1.6.0

# Note: CGAL and Blender need special installation for full test coverage
# CGAL: pip install cgal (may need system dependencies)
# Blender: Automatically downloaded by install.py
# Browser tests also require Chrome/Firefox/Safari to be installed
